Remove sktime dependency
